final commit?

+++ src/main/java/be/ucll/gerecht/repository/GerechtRepository.java
@@ -5,7 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GerechtRepository extends JpaRepository<Gerecht,Long> {
+public interface GerechtRepository extends JpaRepository<Gerecht, Integer> {
     Gerecht findByDescription(String description);
+
     Gerecht findById(int id);
 }
