+++ src/main/java/be/ucll/gerecht/repository/WeekMenuRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WeekMenuRepository extends JpaRepository<WeekMenu,Long> {
+public interface WeekMenuRepository extends JpaRepository<WeekMenu, Integer> {
     WeekMenu findWeekMenuById(int id);
 }
